usb: dwc3: gadget: fix req list in miss isoc event processing
If miss isoc event happens, the current code just move the
current req to the tail of the pending list, and then stop
the active transfer with cmd DWC3_DEPCMD_ENDTRANSFER. After
these operations, the isoc endpoint wait for a XferNotReady
event to restart the transfer with the first pending req
from the pending_list.

If there are already reqs in the pending list when move the
miss isoc reqs to the tail of the pending list, some of the
reqs are out of order.

This patch first moves all of the reqs in the started list
to the tail of the cancelled list, and then move all of the
reqs in the pending list to the tail of the cancelled list.
At last move all of the reqs in the cancelled list to the
pending list. This can make sure all of the reqs are in order
when restart the isoc transfer later.

Signed-off-by: William Wu <william.wu@rock-chips.com>
Change-Id: Idf38d9fd4d483854473c18f792d1996fb5fcab4b