phy: rockchip-naneng-usb2: do apb reset during probe
It's not enough to reset the internal circuit in phy
with the phy power reset signal when probe. It also
needs to reset the apb of phy, with this patch, we
can fix the usb enumeration error on some rv1126/rv1109
products.

Signed-off-by: William Wu <william.wu@rock-chips.com>
Change-Id: Ib6bb0f0aeca7577ad65041a28ede57ba110b7bc0
(cherry picked from commit 01f615bec97bf1174a331714b94f7ac4ab50fd10)